requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-http
apache-airflow-providers-postgres
apache-airflow-providers-google
apache-airflow-providers-amazon
apache-airflow-providers-mysql